+++ src/main.ts
@@ -4,6 +4,7 @@ import {
   disableContextMenu,
   preventNavigation,
   setupFullscreenBtn,
+  setupNotifications,
 } from "./utils/ui/ui";
 import { setupKeyboardControls } from "./utils/controls";
 import { registerServiceWorker } from "./utils/serviceWorkers";
@@ -63,7 +64,7 @@ document.addEventListener("DOMContentLoaded", async () => {
 
   setupScoresDialog();
   setupFullscreenBtn();
-  //setupNotificationsIcon();
+  setupNotifications();
   setupMusic();
   setupMainMenuKeyboardNavigation();
 });