+++ changelog.md
@@ -2,7 +2,8 @@
 
 - Internal:
   - Update some dev dependencies: Tailwindcss, GH Pages, TS ESLint
-  -
+  - Update `FetchRequest.delete` to accept more parameters
+  - Request sent to `DPNS` when/if push subscription revoked by end-user
 
 # 0.8.8.2 Beta Valentine's Edition - 10/02/2026 20:10 UTC+1